+++ src/hardware/cpu/cpu_create.cpp
@@ -15,7 +15,7 @@ void cpu_range(Tensor *A, float min, float step){
     float v=min;
 
     //#pragma omp parallel for
-    for(int i=0; i<A->size; i++){
+    for(unsigned long int i=0; i<A->size; i++){
         A->ptr[i] = v;
         v+=step;
     }
@@ -25,7 +25,7 @@ void cpu_range(Tensor *A, float min, float step){
 void cpu_eye(Tensor *A, int offset){
     _profile(_CPU_EYE, 0);
     #pragma omp parallel for
-    for(int i=0; i<A->size; i++){
+    for(unsigned long int i=0; i<A->size; i++){
         if ((i/A->shape[0]+offset) == i%A->shape[1]){ A->ptr[i] = 1.0f; }  // rows+offset == col?
         else { A->ptr[i] = 0.0f; }
     }
@@ -34,7 +34,7 @@ void cpu_eye(Tensor *A, int offset){
 
 void cpu_diag(Tensor *A, Tensor *B, int k){
     #pragma omp parallel for
-    for(int i=0; i<A->size; i++){
+    for(unsigned long int i=0; i<A->size; i++){
         if ((i/A->shape[0]+k) == i%A->shape[1]){ B->ptr[i] = A->ptr[i]; }  // rows+offset == col?
         else { B->ptr[i] = 0.0f; }
     }
